pyproject.toml
[project]
name = "tooling-trusted-releases"
version = "0.0.1"
description = "Apache Trusted Releases"
authors = [
  { name = "ASF Tooling", email = "dev@tooling.apache.org" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
  "aiofiles>=24.1.0,<25.0.0",
  "aiohttp>=3.11.14",
  "aioshutil (>=1.5,<2.0)",
  "aiosmtplib (>=4.0.0,<5.0.0)",
  "aiosqlite>=0.21.0,<0.22.0",
  "aiozipstream (>=0.4,<0.5)",
  "alembic~=1.14",
  "asfquart @ git+https://github.com/apache/infrastructure-asfquart.git@main",
  "asyncssh>=2.20.0,<3.0.0",
  "blockbuster>=1.5.23,<2.0.0",
  "cmarkgfm>=2024.11.20",
  "cryptography~=44.0",
  "cyclonedx-python-lib[json-validation]>=11.0.0",
  # "dkimpy @ git+https://github.com/sbp/dkimpy.git@main",
  "dnspython>=2.7.0,<3.0.0",
  "dunamai>=1.23.0",
  "email-validator~=2.2.0",
  "gitignore-parser (>=0.1.12,<0.2.0)",
  "greenlet>=3.1.1,<4.0.0",
  "htpy (>=25.7.0,<26.0.0)",
  "hypercorn~=0.17",
  "ldap3 (==2.10.2rc2)",
  "packaging>=25.0",
  "pgpy>=0.6.0",
  "pydantic-xml (>=2.17.2,<3.0.0)",
  "pyjwt (>=2.10.1,<3.0.0)",
  "pynacl>=1.5.0",
  "python-decouple~=3.8",
  "python-gnupg~=0.5",
  "quart-schema[pydantic]~=0.21",
  "quart-wtforms~=1.0.3",
  "rich~=14.0.0",
  "sqlmodel~=0.0.24",
  "standard-imghdr>=3.13.0",
  "yyjson>=4.0.6",
]

[dependency-groups]
dev = [
  "decouple-types>=1.0",
  "djlint>=1.36.4",
  "pre-commit>=2.20.0",
  "pyright>=1.1.393",
  "ruff>=0.9.4",
  "types-aiofiles>=24.1.0.20241221,<25.0.0.0",
]
test = [
  "netifaces>=0.11.0",
  "playwright>=1.51.0",
  "pytest-asyncio>=0.24",
  "pytest>=8.0",
]

# Additional tools

[tool.djlint]
profile = "jinja"
extension = "html"
indent = 2
blank_line_after_tag = "load,extends,include,html"
close_void_tags = true
format_css = true
format_js = true
max_line_length = 120
use_gitignore = true
preserve_blank_lines = true
ignore = "H006,H031"
include = "atr/templates"

[tool.pyright]
include = ["atr", "playwright"]
exclude = [
  ".venv*",
  "**/__pycache__",
  "**/node_modules",
  "tests",
]
ignore = []
defineConstant = { DEBUG = true }
stubPath = "typestubs"
reportMissingImports = "error"
reportMissingTypeStubs = false
# reportUnusedCallResult = "error"
# https://github.com/microsoft/pyright/issues/9579#issuecomment-2549237805
reportUnnecessaryComparison = "error"
reportUnusedCoroutine = "error"
# typeCheckingMode = "strict"
pythonVersion = "3.13"
executionEnvironments = [
  { root = ".", pythonVersion = "3.13" }
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths  = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
extend-exclude = [
    "node_modules",
]

[tool.ruff.lint]
ignore = [
  "S101", # assert
  "S104", # interfaces
  "S106", # password keyword
  "S314", # old expat concerns
  "S603", # subprocess.run
  "S607", # partial path
  "S608", # SQL string query
  "S704", # markupsafe
]
select = [
  "ASYNC", # flake8-async
  "C90",   # cyclomatic-complexity
  "E",     # pycodestyle
  "F",     # pyflakes
  "I",     # isort
  "N",     # pep8-naming
  "RUF",   # ruff-checks
  "S",     # flake8-bandit
  "TC",    # flake8-type-checking
  "TID",   # flake8-tidy-imports
  "UP",    # pyupgrade
  "W"
]

[tool.ruff.lint.per-file-ignores]
"atr/analysis.py" = ["RUF001"]
"atr/db/__init__.py" = ["C901"]
"atr/models/cyclonedx/__init__.py" = ["N815"]
"atr/models/cyclonedx/spdx.py" = ["N815"]
"atr/sbom/conformance.py" = ["C901"]
"atr/sbom/licenses.py" = ["C901"]
"migrations/env.py" = ["E402"]
